requirements.txt
Django>=3.0.1,<=3.1.0
djangorestframework>=3.11.0,<=3.12.0
drf-yasg>=1.17.0,<1.18.0
gunicorn>=20.0.4,<20.1.0
python-logstash>=0.4.6,<0.5.0
django-nose>=1.4.6,<=1.4.9
coverage>=5.0.4,<=5.0.9
factory-boy>=2.12.0,<2.13.0
